# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Books API",
    "version": "1.0.0",
    "description": "A simple API for working with books"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "paths": {
    "/books": {
      "get": {
        "summary": "Get all books",
        "tags": [
          "Books"
        ],
        "responses": {
          "200": {
            "description": "Books returned successfully"
          },
          "500": {
            "description": "Unable to retrieve books"
          }
        }
      },
      "post": {
        "summary": "Create a book",
        "tags": [
          "Books"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id",
                  "authorId",
                  "title",
                  "publicationDate"
                ],
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "authorId": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "publicationDate": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "id": "b4",
                "authorId": "a1",
                "title": "New Book",
                "publicationDate": "2024-01-01"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Book created"
          },
          "400": {
            "description": "Invalid book data"
          }
        }
      }
    },
    "/books/{id}": {
      "get": {
        "summary": "Get one book by id",
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The custom book id, such as b1",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book returned successfully"
          },
          "404": {
            "description": "Book not found"
          },
          "500": {
            "description": "Unable to retrieve book"
          }
        }
      },
      "put": {
        "summary": "Update a book",
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "authorId": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "publicationDate": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Book updated"
          },
          "404": {
            "description": "Book not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a book",
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Book deleted"
          },
          "404": {
            "description": "Book not found"
          }
        }
      }
    },
    "/authors": {
      "get": {
        "summary": "Get all authors",
        "tags": [
          "Authors"
        ],
        "responses": {
          "200": {
            "description": "A list of authors"
          },
          "500": {
            "description": "Unable to retrieve authors"
          }
        }
      },
      "post": {
        "summary": "Create an author",
        "tags": [
          "Authors"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id",
                  "name",
                  "birthYear"
                ],
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "birthYear": {
                    "type": "number"
                  }
                }
              },
              "example": {
                "id": "a4",
                "name": "Example Author",
                "birthYear": 1980
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Author created"
          },
          "400": {
            "description": "Invalid author data"
          }
        }
      }
    },
    "/authors/{id}": {
      "get": {
        "summary": "Get author by ID",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Author returned successfully"
          },
          "404": {
            "description": "Author not found"
          },
          "500": {
            "description": "Unable to retrieve author"
          }
        }
      },
      "put": {
        "summary": "Update an author",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "birthYear": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Author updated"
          },
          "404": {
            "description": "Author not found"
          }
        }
      },
      "delete": {
        "summary": "Delete an author",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Author deleted"
          },
          "404": {
            "description": "Author not found"
          },
          "409": {
            "description": "Author still has books"
          }
        }
      }
    }
  },
  "components": {},
  "tags": []
}